Implementa subscribe dinâmico em todos os tópicos MQTT e criação automática de tópicos Kafka

+++ backend/Iot2Project.Domain/Entities/DeviceData.cs
@@ -4,7 +4,7 @@ public class DeviceData
 {
     public int DeviceId { get; set; }
     public DateTime Timestamp { get; set; }
-    public decimal? Value { get; set; }
+    public float? Value { get; set; }
     public int? UserId { get; set; }
     public string? Command { get; set; }
 }
